chore: move utils/proto into internal/proto
进一步将目录结构与MiraiGo对齐
