+++ internal/proto/dynamic.go
