+++ internal/proto/dynamic_test.go
