+++ internal/proto/wrapper.go
