+++ message/elements.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/LagrangeDev/LagrangeGo/utils"
 
+	"github.com/LagrangeDev/LagrangeGo/internal/proto"
 	"github.com/LagrangeDev/LagrangeGo/packets/pb/message"
-	"github.com/LagrangeDev/LagrangeGo/utils/proto"
 )
 
 var messageLogger = utils.GetLogger("message")
