+++ packets/oidb/GroupSetReaction.go
@@ -1,8 +1,8 @@
 package oidb
 
 import (
+	"github.com/LagrangeDev/LagrangeGo/internal/proto"
 	"github.com/LagrangeDev/LagrangeGo/packets/pb/service/oidb"
-	"github.com/LagrangeDev/LagrangeGo/utils/proto"
 )
 
 func BuildGroupSetReactionReq(groupUin, sequence uint32, code string) (*OidbPacket, error) {
