+++ packets/tlv/qrcode.go
@@ -1,8 +1,8 @@
 package tlv
 
 import (
+	"github.com/LagrangeDev/LagrangeGo/internal/proto"
 	"github.com/LagrangeDev/LagrangeGo/utils/binary"
-	"github.com/LagrangeDev/LagrangeGo/utils/proto"
 )
 
 func T11(unusualSign []byte) []byte {
