--- utils/proto/doc.go
@@ -1,3 +1,0 @@
-package proto
-
-// from https://github.com/Mrs4s/MiraiGo/tree/master/internal/proto
